Move run server unit tests into own sub directory
And use babel-plugin-module-resolver to alias lib modules
