+++ package.json
@@ -48,6 +48,7 @@
     "@bigtest/meta": "bigtestjs/meta",
     "babel-eslint": "^9.0.0-beta.0",
     "babel-loader": "^8.0.0-beta.0",
+    "babel-plugin-module-resolver": "^3.1.1",
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
     "css-loader": "^1.0.0",
