+++ tests/unit/run/process-test.js
@@ -2,7 +2,7 @@ import { describe, beforeEach, it } from 'mocha';
 import { WritableStream } from 'memory-streams';
 import { expect, when } from '../helpers';
 
-import Process from '../../../lib/run/process';
+import Process from '@run/process';
 
 describe('Unit: Process', () => {
   it('can be created with specific properties', () => {
