+++ tests/unit/run/servers/client-test.js
@@ -1,8 +1,8 @@
 import { describe, beforeEach, afterEach, it } from 'mocha';
-import { expect, request } from '../helpers';
+import { expect, request } from '../../helpers';
 
-import WebServer from '../../../lib/run/servers/web';
-import ClientServer from '../../../lib/run/servers/client';
+import WebServer from '@run/servers/web';
+import ClientServer from '@run/servers/client';
 
 describe('Unit: ClientServer', () => {
   let test;
